handle claiming resource with required description for set and singleton types
