+++ packages/frinx-uniresource-ui/src/pages/pool-detail-page/claim-resource-modal/claim-resource-modal.tsx
@@ -2,6 +2,7 @@ import React, { FC } from 'react';
 import ClaimResourceAllocIpv4PrefixModal from './claim-resource-allocating-modals/claim-resource-allocating-ipv4_prefix-modal';
 import ClaimResourceAllocIpv6PrefixModal from './claim-resource-allocating-modals/claim-resource-allocating-ipv6_prefix';
 import ClaimResourceAllocatingVlanRangeModal from './claim-resource-allocating-modals/claim-resource-allocationg-vlan_range-modal';
+import ClaimResourceDefaultModal from './claim-resource-allocating-modals/claim-resource-default-modal';
 
 // eslint-disable-next-line no-shadow
 enum ClaimModalVariant {
@@ -24,6 +25,9 @@ const ClaimResourceModal: FC<Props> = ({ variant, ...props }) => {
       {variant === ClaimModalVariant.IPV4_PREFIX && <ClaimResourceAllocIpv4PrefixModal {...props} />}
       {variant === ClaimModalVariant.IPV6_PREFIX && <ClaimResourceAllocIpv6PrefixModal {...props} />}
       {variant === ClaimModalVariant.VLAN_RANGE && <ClaimResourceAllocatingVlanRangeModal {...props} />}
+      {variant !== ClaimModalVariant.IPV4_PREFIX &&
+        variant !== ClaimModalVariant.IPV6_PREFIX &&
+        variant !== ClaimModalVariant.VLAN_RANGE && <ClaimResourceDefaultModal {...props} />}
     </>
   );
 };
